Fix studio module import paths after JS tree move
